apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
python-dotenv==1.0.1
PyJWT[crypto]==2.9.0
httpx==0.27.2
sqlalchemy==2.0.35
alembic==1.13.3
psycopg[binary]==3.2.4
requests==2.31.0
